将被下载的文件,上传阿里云OSS的时候进行压缩,节省空间和流量

+++ udf-starter-file-aliyun-oss-api/src/main/java/org/itkk/udf/file/aliyun/oss/api/download/DownLoadProcessAspect.java
@@ -11,7 +11,9 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.itkk.udf.cache.redis.CacheRedisProperties;
+import org.itkk.udf.core.ApplicationConfig;
 import org.itkk.udf.core.exception.SystemRuntimeException;
+import org.itkk.udf.core.utils.ZipUtil;
 import org.itkk.udf.file.aliyun.oss.api.OssWarpper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.annotation.Order;
@@ -19,7 +21,9 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
+import java.util.Date;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -41,7 +45,7 @@ public class DownLoadProcessAspect {
     /**
      * DOWNLOAD_WAITING_CACHE_EXPIRATION(待执行,执行中,缓存过期时间为1天)
      */
-    public static final long DOWNLOAD_WAITING_CACHE_EXPIRATION = 60 * 24;
+    public static final long DOWNLOAD_WAITING_CACHE_EXPIRATION = 60L * 24L;
 
     /**
      * DOWNLOAD_CACHE_EXPIRATION(执行完成,执行错误,缓存过期时间为10分钟)
@@ -67,6 +71,12 @@ public class DownLoadProcessAspect {
     private OssWarpper ossWarpper;
 
     /**
+     * applicationConfig
+     */
+    @Autowired
+    private ApplicationConfig applicationConfig;
+
+    /**
      * around
      *
      * @param proceedingJoinPoint proceedingJoinPoint
@@ -87,29 +97,56 @@ public class DownLoadProcessAspect {
             //构造对象
             DownloadInfo info = new DownloadInfo();
             info.setId(param.getId());
+            //定义文件变量
+            File file = null;
+            File zipFile = null;
+            //开始处理
             try {
                 //更新缓存
                 info.setStatus(DownConstant.DOWNLOAD_PROCESS_STATUS.STATUS_2.value());
+                info.setProcessStartDate(new Date());
                 redisTemplate.opsForValue().set(key, info, DOWNLOAD_WAITING_CACHE_EXPIRATION, TimeUnit.MINUTES);
                 //执行
                 result = proceedingJoinPoint.proceed();
                 //获得本地文件
-                File file = (File) result;
+                file = (File) result;
+                //文件zip压缩
+                ZipUtil zipUtil = new ZipUtil(applicationConfig.getEncoding());
+                String zipFilePath = DownConstant.tempdir() + param.getId() + ".zip";
+                File[] files = {file};
+                zipUtil.compressFiles2Zip(files, zipFilePath);
+                zipFile = new File(zipFilePath);
                 //上传文件到阿里云OSS
-                String objectKey = ossWarpper.uploadFile(param.getOssCode(), file);
-                //删除本地文件
-                Files.delete(file.toPath());
+                String objectKey = ossWarpper.uploadFile(param.getOssCode(), zipFile);
                 //更新缓存
                 info.setStatus(DownConstant.DOWNLOAD_PROCESS_STATUS.STATUS_3.value());
+                info.setProcessEndDate(new Date());
                 info.setObjectKey(objectKey);
                 redisTemplate.opsForValue().set(key, info, DOWNLOAD_CACHE_EXPIRATION, TimeUnit.MINUTES);
             } catch (Throwable e) {
                 //更新缓存
                 info.setStatus(DownConstant.DOWNLOAD_PROCESS_STATUS.STATUS_4.value());
+                info.setProcessEndDate(new Date());
                 info.setErrorMsg(e.getMessage());
                 redisTemplate.opsForValue().set(key, info, DOWNLOAD_CACHE_EXPIRATION, TimeUnit.MINUTES);
                 //抛出异常
                 throw new SystemRuntimeException(e);
+            } finally {
+                //删除本地文件
+                if (file != null) {
+                    try {
+                        Files.delete(file.toPath());
+                    } catch (IOException e) {
+                        log.error("error:", e);
+                    }
+                }
+                if (zipFile != null) {
+                    try {
+                        Files.delete(zipFile.toPath());
+                    } catch (IOException e) {
+                        log.error("error:", e);
+                    }
+                }
             }
         }
         //返回
